python/requirements.txt
numpy
pandas
scikit-learn
umap-learn
plotly
ijson
tqdm